# server/docs/swagger.yaml
definitions:
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  request.InitDB:
    properties:
      adminPassword:
        type: string
      dbName:
        description: 数据库名
        type: string
      dbPath:
        description: sqlite数据库文件路径
        type: string
      dbType:
        description: 数据库类型
        type: string
      host:
        description: 服务器地址
        type: string
      password:
        description: 数据库密码
        type: string
      port:
        description: 数据库连接端口
        type: string
      userName:
        description: 数据库用户名
        type: string
    required:
    - adminPassword
    - dbName
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  system.Condition:
    properties:
      ID:
        description: 主键ID
        type: integer
      column:
        type: string
      createdAt:
        description: 创建时间
        type: string
      from:
        type: string
      operator:
        type: string
      templateID:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  system.JoinTemplate:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      joins:
        type: string
      "on":
        type: string
      table:
        type: string
      templateID:
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  system.SysExportTemplate:
    properties:
      ID:
        description: 主键ID
        type: integer
      conditions:
        items:
          $ref: '#/definitions/system.Condition'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dbName:
        description: 数据库名称
        type: string
      joinTemplate:
        items:
          $ref: '#/definitions/system.JoinTemplate'
        type: array
      limit:
        type: integer
      name:
        description: 模板名称
        type: string
      order:
        type: string
      tableName:
        description: 表名称
        type: string
      templateID:
        description: 模板标识
        type: string
      templateInfo:
        description: 模板信息
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.5
paths:
...
  /init/initdb:
    post:
      parameters:
      - description: 初始化数据库参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.InitDB'
      produces:
      - application/json
      responses:
        "200":
          description: 初始化用户数据库
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 初始化用户数据库
      tags:
      - InitDB
...
  /sysExportTemplate/createSysExportTemplate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建导出模板
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysExportTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建导出模板
      tags:
      - SysExportTemplate
  /sysExportTemplate/deleteSysExportTemplate:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除导出模板
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysExportTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除导出模板
      tags:
      - SysExportTemplate
  /sysExportTemplate/deleteSysExportTemplateByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除导出模板
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除导出模板
      tags:
      - SysExportTemplate
  /sysExportTemplate/exportExcel:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 导出表格模板
      tags:
      - SysExportTemplate
  /sysExportTemplate/findSysExportTemplate:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 数据库名称
        in: query
        name: dbName
        type: string
      - in: query
        name: limit
        type: integer
      - description: 模板名称
        in: query
        name: name
        type: string
      - in: query
        name: order
        type: string
      - description: 表名称
        in: query
        name: tableName
        type: string
      - description: 模板标识
        in: query
        name: templateID
        type: string
      - description: 模板信息
        in: query
        name: templateInfo
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询导出模板
      tags:
      - SysExportTemplate
  /sysExportTemplate/getSysExportTemplateList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 数据库名称
        in: query
        name: dbName
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - description: 模板名称
        in: query
        name: name
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      - description: 表名称
        in: query
        name: tableName
        type: string
      - description: 模板标识
        in: query
        name: templateID
        type: string
      - description: 模板信息
        in: query
        name: templateInfo
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取导出模板列表
      tags:
      - SysExportTemplate
  /sysExportTemplate/importExcel:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 导入表格
      tags:
      - SysImportTemplate
  /sysExportTemplate/updateSysExportTemplate:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新导出模板
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.SysExportTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新导出模板
      tags:
      - SysExportTemplate
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"